fix startup
