--- bin/process-engine.js
@@ -1,50 +1,0 @@
-const InvocationContainer = require('addict-ioc').InvocationContainer;
-const logger = require('loggerhythm').Logger.createLogger('bootstrapper');
-
-const iocModuleNames = [
-  '@essential-projects/bootstrapper',
-  '@essential-projects/bootstrapper_node',
-  '@essential-projects/event_aggregator',
-  '@essential-projects/http_extension',
-  '@essential-projects/services',
-  '@process-engine/consumer_api_core', // Required by the process engine's UserTask handler
-  '@process-engine/flow_node_instance.repository.sequelize',
-  '@process-engine/iam',
-  '@process-engine/management_api_core',
-  '@process-engine/management_api_http',
-  '@process-engine/deployment_api_core',
-  '@process-engine/deployment_api_http',
-  '@process-engine/process_engine',
-  '@process-engine/process_model.repository.sequelize',
-  '@process-engine/timers.repository.sequelize',
-];
-
-const iocModules = iocModuleNames.map((moduleName) => {
-  return require(`${moduleName}/ioc_module`);
-});
-
-async function start() {
-
-  const container = new InvocationContainer({
-    defaults: {
-      conventionCalls: ['initialize'],
-    },
-  });
-
-  for (const iocModule of iocModules) {
-    iocModule.registerInContainer(container);
-  }
-
-  container.validateDependencies();
-
-  try {
-    const bootstrapper = await container.resolveAsync('AppBootstrapper');
-    await bootstrapper.start();
-
-    logger.info('Bootstrapper started successfully.');
-  } catch (error) {
-    logger.error('Bootstrapper failed to start.', error);
-  }
-}
-
-start();
